Use util functions in table1
Table1 code was copied from another project. Use local utility functions
where possible.
